Move cost & constraint ingredients to separate files.
